src/main/java/com/github/mc1arke/sonarqube/plugin/ce/pullrequest/markup/Link.java
/*
 * Copyright (C) 2020 Markus Heberling
 *
 * This program is free software; you can redistribute it and/or
 * modify it under the terms of the GNU Lesser General Public
 * License as published by the Free Software Foundation; either
 * version 3 of the License, or (at your option) any later version.
 *
 * This program is distributed in the hope that it will be useful,
 * but WITHOUT ANY WARRANTY; without even the implied warranty of
 * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
 * Lesser General Public License for more details.
 *
 * You should have received a copy of the GNU Lesser General Public License
 * along with this program; if not, write to the Free Software Foundation,
 * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 *
 */
package com.github.mc1arke.sonarqube.plugin.ce.pullrequest.markup;

public final class Link extends Node {

    private final String url;

    public Link(String url, Node... children) {
        super(children);
        this.url=url;
      }

    public String getUrl() {
        return url;
    }

    @Override
    boolean isValidChild(Node child) {
        return child instanceof Text;
    }
}
